Change getExamMaxPoints method return value from long to int

+++ src/main/java/pl/archala/testme/controllers/ExamController.java
@@ -37,7 +37,7 @@ public class ExamController {
     }
 
     @PostMapping("/exam/{id}")
-    public long getExamMaxPoints(@PathVariable("id") Long id) {
+    public int getExamMaxPoints(@PathVariable("id") Long id) {
         return examService.getMaxPossibleExamPoints(id);
     }
